rename "autoreduce" directory to "mr_autoreduce"
same for directory "livereduce"
