+++ Makefile
@@ -10,20 +10,19 @@ base:
 	cp -R src/mr_reduction/*.py $(prefix)/autoreduce/mr_reduction
 
 autoreduce: base
-	cp -R src/autoreduce/*.template $(prefix)/autoreduce
-	cp -R src/livereduce/*.py $(prefix)/livereduce
+	cp -R src/mr_autoreduce/*.template $(prefix)/autoreduce
+	cp -R src/mr_livereduce/*.py $(prefix)/livereduce
 
 test:
 	python tests/unit_tests.py
 
 check:
 	diff --exclude="*.pyc" -r src/mr_reduction $(prefix)/autoreduce/mr_reduction
-	diff --exclude="*.pyc" --exclude="livereduce.conf" -r -q src/livereduce $(prefix)/livereduce
-	diff --exclude="*.pyc" -r -q src/autoreduce/*.template $(prefix)/autoreduce
+	diff --exclude="*.pyc" --exclude="livereduce.conf" -r -q src/mr_livereduce $(prefix)/livereduce
+	diff --exclude="*.pyc" -r -q src/mr_autoreduce/*.template $(prefix)/autoreduce
 
 .PHONY: install
 .PHONY: base
 .PHONY: autoreduce
 .PHONY: test
 .PHONY: check
-
