--- src/livereduce/reduce_REF_M_live_proc.py
@@ -1,12 +1,0 @@
-"""
-    Accumulation workspace
-
-    To test: nc -l  31466 < REF_M_25640.adara
-"""
-from mantid import simpleapi
-
-simpleapi.logger.notice("Starting proc")
-try:
-    simpleapi.CloneWorkspace(InputWorkspace=input, OutputWorkspace=output)
-except:
-    return input
