+++ src/mr_autoreduce/reduce_REF_M.py.template
