+++ src/mr_autoreduce/templates/reduction_options.html
@@ -174,8 +174,8 @@ if you want to use the region on each side of the peak to estimate your backgrou
             </div>
         </div>
 
-        <button id="submit_form_button" style="padding: 10px 20px; font-size: 22px;" onclick="submitForm()">Submit</button>
-        <button id="view_report_button" style="padding: 10px 20px; font-size: 22px;" onclick="viewReport()" disabled>View Report</button>
+        <button id="submit_form_button" title="Reduction may take up to two minutes" style="padding: 10px 20px; font-size: 22px;" onclick="submitForm()">Submit</button>
+        <button id="view_report_button" title="An HTML report will be generated after reduction is finished" style="padding: 10px 20px; font-size: 22px;" onclick="viewReport()" disabled>View Report</button>
     </div>
 
     <script>
