+++ tests/unit/mr_autoreduce/test_template.py
